Rename `database/api` to `database/index`
